requirements.txt
Flask==2.0.1
psycopg2-binary==2.9.1
Werkzeug==2.0.1
python-dotenv==0.19.0
azure-storage-blob==12.8.0
WTForms==3.0.0
Flask-WTF==1.0.0
gunicorn==20.1.0